+++ ReSharper.FSharp/src/FSharp.Psi/src/Impl/GeneratedMembersUtil.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using JetBrains.Annotations;
 using JetBrains.ReSharper.Plugins.FSharp.Psi.Impl.Cache2;
 using JetBrains.ReSharper.Plugins.FSharp.Psi.Impl.Cache2.Parts;
 using JetBrains.ReSharper.Plugins.FSharp.Psi.Impl.DeclaredElement.CompilerGenerated;
@@ -81,6 +82,27 @@ namespace JetBrains.ReSharper.Plugins.FSharp.Psi.Impl
       }
 
       return result.ResultingList();
+    }
+
+    public static IEnumerable<IDeclaredElement> GetUnionCaseGeneratedMembers([NotNull] this IUnionCase unionCase)
+    {
+      var type = unionCase.GetContainingType();
+      if (type == null)
+        return EmptyList<IDeclaredElement>.Instance;
+
+      var result = new List<IDeclaredElement>();
+      foreach (var member in type.GetMembers())
+      {
+        if (member is IFSharpGeneratedFromOtherElement generated && unionCase.Equals(generated.OriginElement))
+          result.Add(member);
+
+        if (member is FSharpUnionTagsClass tagsClass)
+          foreach (var tagsMember in tagsClass.GetMembers())
+            if (tagsMember is IFSharpGeneratedFromOtherElement tag && unionCase.Equals(tag.OriginElement))
+              result.Add(tagsMember);
+      }
+
+      return result;
     }
   }
 }